fix(modeling): align dynamic bias dtype and clarify ES decoupled guard

+++ src/deberta/modeling/deberta_v2_native.py
@@ -381,9 +381,10 @@ class DisentangledSelfAttention(nn.Module):
         pos_query_layer: torch.Tensor | None = None
 
         if ("c2p" in self.pos_att_type) or ("p2p" in self.pos_att_type):
-            pos_key_layer = self._project_rel(rel_embeddings, use_query=False)
+            # Keep dynamic path on the same explicit dtype contract as cached_bmm/stable.
+            pos_key_layer = self._project_rel(rel_embeddings, use_query=False).to(dtype=query_layer.dtype)
         if ("p2c" in self.pos_att_type) or ("p2p" in self.pos_att_type):
-            pos_query_layer = self._project_rel(rel_embeddings, use_query=True)
+            pos_query_layer = self._project_rel(rel_embeddings, use_query=True).to(dtype=query_layer.dtype)
 
         if "c2p" in self.pos_att_type:
             if pos_key_layer is None:
